+++ Game/DungeonFloor.py
@@ -51,15 +51,18 @@ class DungeonFloor:
         random_y = 1
         random_x = 0
         
-        npc0 = NPC( self.game, y=1, x=0 )   
-        npc1 = NPC( self.game, y=2, x=0 )   
-        npc2 = NPC( self.game, y=3, x=0 )   
+        #npc0 = NPC( self.game, y=1, x=0 )   
+        #npc1 = NPC( self.game, y=2, x=0 )   
+        #npc2 = NPC( self.game, y=3, x=0 )   
         
-        self.npcs = [ npc0, npc1, npc2 ]
+        #self.npcs = [ npc0, npc1, npc2 ]
+        self.npcs = [  ]
 
+        # experimenting with 2 items 1 tile
         item0 = Item( "Short Sword", itemclass=ItemClass.WEAPON, y=4, x=0, weight=1 )
+        item1 = Item( "Long Sword", itemclass=ItemClass.WEAPON, y=4, x=0, weight=1 )
 
-        self.items = [ item0 ] 
+        self.items = [ item0, item1 ]